Seperate trace and logs into different classes.
