+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/src/MessageData.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using OpenTelemetry.Logs;
+
+namespace Azure.Monitor.OpenTelemetry.Exporter.Models
+{
+    internal partial class MessageData
+    {
+        public MessageData(int version, LogRecord logRecord) : base(version)
+        {
+            Message = logRecord.State.ToString();
+            SeverityLevel = LogsHelper.GetSeverityLevel(logRecord.LogLevel);
+        }
+    }
+}
