+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/tests/Azure.Monitor.OpenTelemetry.Exporter.Tests/RemoteDependencyDataTests.cs
@@ -1,0 +1,142 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.Diagnostics;
+using System.Globalization;
+using Azure.Monitor.OpenTelemetry.Exporter.Models;
+using OpenTelemetry.Trace;
+using Xunit;
+
+namespace Azure.Monitor.OpenTelemetry.Exporter
+{
+    public class RemoteDependencyDataTests : ActivityListenerInitializer
+    {
+        private const string ActivitySourceName = "RemoteDependencyDataTests";
+        private const string ActivityName = "RemoteDependencyDataActivity";
+
+        [Theory]
+        [InlineData("mssql")]
+        [InlineData("redis")]
+        public void ValidateDBDependencyType(string dbSystem)
+        {
+            using ActivitySource activitySource = new ActivitySource(ActivitySourceName);
+            using var activity = activitySource.StartActivity(
+                ActivityName,
+                ActivityKind.Server,
+                parentContext: new ActivityContext(ActivityTraceId.CreateRandom(), ActivitySpanId.CreateRandom(), ActivityTraceFlags.Recorded),
+                startTime: DateTime.UtcNow);
+
+            activity.SetTag(SemanticConventions.AttributeDbSystem, dbSystem);
+
+            var monitorTags = TraceHelper.EnumerateActivityTags(activity);
+
+            var remoteDependencyDataType = new RemoteDependencyData(2, activity, ref monitorTags).Type;
+            var expectedType = RemoteDependencyData.SqlDbs.Contains(dbSystem) ? "SQL" : dbSystem;
+
+            Assert.Equal(expectedType, remoteDependencyDataType);
+        }
+
+        [Fact]
+        public void DependencyTypeisSetToInProcForInternalSpanWithParent()
+        {
+            using ActivitySource activitySource = new ActivitySource(ActivitySourceName);
+            using var parentActivity = activitySource.StartActivity("ParentActivity", ActivityKind.Internal);
+            using var childActivity = activitySource.StartActivity("ChildActivity", ActivityKind.Internal);
+
+            var monitorTagsParent = TraceHelper.EnumerateActivityTags(parentActivity);
+
+            var remoteDependencyDataTypeForParent = new RemoteDependencyData(2, parentActivity, ref monitorTagsParent).Type;
+
+            Assert.Null(remoteDependencyDataTypeForParent);
+
+            var monitorTagsChild = TraceHelper.EnumerateActivityTags(childActivity);
+
+            var remoteDependencyDataTypeForChild = new RemoteDependencyData(2, childActivity, ref monitorTagsChild).Type;
+
+            Assert.Equal("InProc", remoteDependencyDataTypeForChild);
+        }
+
+        [Fact]
+        public void ValidateHttpRemoteDependencyData()
+        {
+            using ActivitySource activitySource = new ActivitySource(ActivitySourceName);
+            using var activity = activitySource.StartActivity(
+                ActivityName,
+                ActivityKind.Client,
+                parentContext: new ActivityContext(ActivityTraceId.CreateRandom(), ActivitySpanId.CreateRandom(), ActivityTraceFlags.Recorded),
+                startTime: DateTime.UtcNow);
+            activity.Stop();
+
+            var httpUrl = "https://www.foo.bar/search";
+            activity.SetStatus(Status.Ok);
+            activity.SetTag(SemanticConventions.AttributeHttpMethod, "GET");
+            activity.SetTag(SemanticConventions.AttributeHttpUrl, httpUrl); // only adding test via http.url. all possible combinations are covered in HttpHelperTests.
+            activity.SetTag(SemanticConventions.AttributeHttpStatusCode, null);
+
+            var monitorTags = TraceHelper.EnumerateActivityTags(activity);
+
+            var remoteDependencyData = new RemoteDependencyData(2, activity, ref monitorTags);
+
+            Assert.Equal("GET /search", remoteDependencyData.Name);
+            Assert.Equal(activity.Context.SpanId.ToHexString(), remoteDependencyData.Id);
+            Assert.Equal(httpUrl, remoteDependencyData.Data);
+            Assert.Equal("0", remoteDependencyData.ResultCode);
+            Assert.Equal(activity.Duration.ToString("c", CultureInfo.InvariantCulture), remoteDependencyData.Duration);
+            Assert.Equal(activity.GetStatus() != Status.Error, remoteDependencyData.Success);
+            Assert.True(remoteDependencyData.Properties.Count == 0);
+            Assert.True(remoteDependencyData.Measurements.Count == 0);
+        }
+
+        [Fact]
+        public void ValidateDbRemoteDependencyData()
+        {
+            using ActivitySource activitySource = new ActivitySource(ActivitySourceName);
+            using var activity = activitySource.StartActivity(
+                ActivityName,
+                ActivityKind.Client,
+                parentContext: new ActivityContext(ActivityTraceId.CreateRandom(), ActivitySpanId.CreateRandom(), ActivityTraceFlags.Recorded),
+                startTime: DateTime.UtcNow);
+            activity.Stop();
+
+            activity.SetStatus(Status.Ok);
+            activity.SetTag(SemanticConventions.AttributeDbSystem, "mssql");
+            activity.SetTag(SemanticConventions.AttributePeerService, "localhost"); // only adding test via peer.service. all possible combinations are covered in HttpHelperTests.
+            activity.SetTag(SemanticConventions.AttributeDbStatement, "Select * from table");
+
+            var monitorTags = TraceHelper.EnumerateActivityTags(activity);
+
+            var remoteDependencyData = new RemoteDependencyData(2, activity, ref monitorTags);
+
+            Assert.Equal(ActivityName, remoteDependencyData.Name);
+            Assert.Equal(activity.Context.SpanId.ToHexString(), remoteDependencyData.Id);
+            Assert.Equal("Select * from table", remoteDependencyData.Data);
+            Assert.Null(remoteDependencyData.ResultCode);
+            Assert.Equal(activity.Duration.ToString("c", CultureInfo.InvariantCulture), remoteDependencyData.Duration);
+            Assert.Equal(activity.GetStatus() != Status.Error, remoteDependencyData.Success);
+            Assert.True(remoteDependencyData.Properties.Count == 0);
+            Assert.True(remoteDependencyData.Measurements.Count == 0);
+        }
+
+        [Fact]
+        public void HttpDependencyNameIsActivityDisplayNameByDefault()
+        {
+            using ActivitySource activitySource = new ActivitySource(ActivitySourceName);
+            using var activity = activitySource.StartActivity(
+                ActivityName,
+                ActivityKind.Client,
+                parentContext: new ActivityContext(ActivityTraceId.CreateRandom(), ActivitySpanId.CreateRandom(), ActivityTraceFlags.Recorded),
+                startTime: DateTime.UtcNow);
+
+            activity.SetTag(SemanticConventions.AttributeHttpMethod, "GET");
+
+            activity.DisplayName = "HTTP GET";
+
+            var monitorTags = TraceHelper.EnumerateActivityTags(activity);
+
+            var remoteDependencyDataName = new RemoteDependencyData(2, activity, ref monitorTags).Name;
+
+            Assert.Equal(activity.DisplayName, remoteDependencyDataName);
+        }
+    }
+}
