+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/tests/Azure.Monitor.OpenTelemetry.Exporter.Tests/RequestDataTests.cs
@@ -1,0 +1,74 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.Diagnostics;
+using System.Globalization;
+using Azure.Monitor.OpenTelemetry.Exporter.Models;
+using OpenTelemetry.Trace;
+using Xunit;
+
+namespace Azure.Monitor.OpenTelemetry.Exporter
+{
+    public class RequestDataTests : ActivityListenerInitializer
+    {
+        private const string ActivitySourceName = "RequestDataTests";
+        private const string ActivityName = "RequestDataActivity";
+
+        [Fact]
+        public void ValidateHttpRequestData()
+        {
+            using ActivitySource activitySource = new ActivitySource(ActivitySourceName);
+            using var activity = activitySource.StartActivity(
+                ActivityName,
+                ActivityKind.Server,
+                parentContext: new ActivityContext(ActivityTraceId.CreateRandom(), ActivitySpanId.CreateRandom(), ActivityTraceFlags.Recorded),
+                startTime: DateTime.UtcNow);
+            activity.Stop();
+
+            var httpUrl = "https://www.foo.bar/search";
+            activity.SetStatus(Status.Ok);
+            activity.SetTag(SemanticConventions.AttributeHttpMethod, "GET");
+            activity.SetTag(SemanticConventions.AttributeHttpRoute, "/search");
+            activity.SetTag(SemanticConventions.AttributeHttpUrl, httpUrl); // only adding test via http.url. all possible combinations are covered in HttpHelperTests.
+            activity.SetTag(SemanticConventions.AttributeHttpStatusCode, null);
+
+            var monitorTags = TraceHelper.EnumerateActivityTags(activity);
+
+            var requestData = new RequestData(2, activity, ref monitorTags);
+
+            Assert.Equal("GET /search", requestData.Name);
+            Assert.Equal(activity.Context.SpanId.ToHexString(), requestData.Id);
+            Assert.Equal(httpUrl, requestData.Url);
+            Assert.Equal("0", requestData.ResponseCode);
+            Assert.Equal(activity.Duration.ToString("c", CultureInfo.InvariantCulture), requestData.Duration);
+            Assert.Equal(activity.GetStatus() != Status.Error, requestData.Success);
+            Assert.Null(requestData.Source);
+            Assert.True(requestData.Properties.Count == 0);
+            Assert.True(requestData.Measurements.Count == 0);
+        }
+
+        [Theory]
+        [InlineData("200")]
+        [InlineData(null)]
+        public void ValidateHttpRequestDataResponseCode(string httpStatusCode)
+        {
+            using ActivitySource activitySource = new ActivitySource(ActivitySourceName);
+            using var activity = activitySource.StartActivity(
+                ActivityName,
+                ActivityKind.Server,
+                parentContext: new ActivityContext(ActivityTraceId.CreateRandom(), ActivitySpanId.CreateRandom(), ActivityTraceFlags.Recorded),
+                startTime: DateTime.UtcNow);
+
+            var httpResponseCode = httpStatusCode ?? "0";
+            activity.SetTag(SemanticConventions.AttributeHttpUrl, "https://www.foo.bar/search");
+            activity.SetTag(SemanticConventions.AttributeHttpStatusCode, httpStatusCode);
+
+            var monitorTags = TraceHelper.EnumerateActivityTags(activity);
+
+            var requestData = new RequestData(2, activity, ref monitorTags);
+
+            Assert.Equal(httpResponseCode, requestData.ResponseCode);
+        }
+    }
+}
